+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Player Rank Hider works now with all messages.
 + Add two more chat filter categories: Powder Mining and Winter Gifts.
 + Add catacombs class level color to party finder.
++ Add wishing compass uses amount display.
 
 ### Removals
 - Removed additional settings for the chat design, like channel prefix and skyblock level